+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 Features:
 - Responsive, adaptive size
 - Dynamic load size data from predefined files from assets
+
+# [1.0.1] - 2022/02/15
+## Update version
+CHANGES:
+- Update path version to 1.8.0
+- Throw exception if not found value for (string) key
+- Fix generate size data key class